Run `cargo clippy --fix` on the repo

See merge request crypto/swanky!210

+++ mac-n-cheese/wire-map/benches/wire-map.rs
@@ -16,7 +16,7 @@ fn main() {
     for base in bases {
         wm.alloc(base, len).unwrap();
     }
-    let len = len as u64;
+    let len = len;
     for i in 0..len {
         for base in bases {
             wm.insert(base + i, black_box(base + i));
